requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
twilio==8.11.1
urllib3==2.1.0
yarl==1.9.4
